Add conditions and transitions
An array of conditions could be provided to config. Only when all conditions are
met, the actions will be applied. This adds the following types of conditions:
- lastHumanActivity: condition would be true if last human activity happened before
  provided days
- lastBotActivity: condition would be true if last bot activity happened before
  provided days
- noRecentCommentTags: condition would be true if no tag in the array was a recent
  comment tag
- recentCommentTag: check's if the most recent comment has the provided tag

In addition to actions, transition could also be provided in config.
Transition would be applied when all conditions pass. This adds two types of
transitions: close and resolve.

This diff also adds ability to add a commentTag to each comment. Provided
commentTag would be added as #<commentTag> to the end of the provided comment.

- mode: if mode is set to 'dryrun', the bot would run by outputting log lines
  and doesn't issue any JIRA API.
- bots: an array that stores all the bot names. these botnames are used to
  detect lastBotActivity
- transitionIds: a hash with transition name and corresponding jira transition
  id. Without transitionIds, transition would fail.