Add conditions and transitions
An array of conditions could be provided to config. Only when all conditions are
met, the actions will be applied. This adds the following types of conditions:
- lastHumanActivity: condition would be true if last human activity happened before
  provided days
- recentCommentTag: check's if the most recent comment has the provided tag
- resolution: check's if the resolution happened before provided days

In addition to actions, transition could also be provided in config.
Transition would be applied when all conditions pass. This adds two types of
transitions: close and resolve.

This diff also adds ability to add a commentTag to each comment. Provided
commentTag would be added as #<commentTag> to the end of the provided comment.